Fix bug in subgraph_by_nodes
This commit fixes a bug in the implementation of subgraph_by_nodes which
was causing an out-out-bounds index error. The problem was that the
previous implementation wasn't correctly manipulating the edge_ids and
nodes_ids internal maps we use to perform the bookeeping between domain
model ids and petgraph ids, and this was the culprit. Using a fresh, new
implementation that correctly handles that, fixed the problem.